+++ include/number_theory/cipher.hpp
@@ -15,7 +15,7 @@ class Cipher {
 
     public:
         /* compute the caesar cipher of a given string with a shift key */
-        std::string caesar(std::string plaintext, int key); 
+        std::string caesar(std::string plaintext, int64_t key); 
         /* encode and prepare the given string for the cipher */
         std::string keyword_encode(std::string key);
         /* compute the keyword cipher of a given string */
