+++ modules/calculus/deriv.cpp
@@ -10,7 +10,7 @@
 #include "../../include/calculus/deriv.hpp"
 
 /*
-long long Calculus::derivativeTerm(std::string p_term, long long val) {
+int64_t Calculus::derivativeTerm(std::string p_term, long long val) {
     // Get coefficient
     std::string coeffStr = "";
     int i;
@@ -18,7 +18,7 @@ long long Calculus::derivativeTerm(std::string p_term, long long val) {
     for (i = 0; p_term[i] != 'x'; i++) {
         coeffStr.push_back(p_term[i]);
     }
-    long long coeff = atol(coeffStr.c_str());
+    int64_t coeff = atol(coeffStr.c_str());
  
     // Get Power (Skip 2 characters for x and ^)
     std::string powStr = "";
@@ -26,27 +26,27 @@ long long Calculus::derivativeTerm(std::string p_term, long long val) {
     for (i = i + 2; i != p_term.size(); i++) {
         powStr.push_back(p_term[i]);
     }
-    long long expo = atol(powStr.c_str());
+    int64_t expo = atol(powStr.c_str());
  
     // For ax^n, we return anx^(n-1)
 
-    long long result = coeff * expo * pow(val, expo - 1);
+    int64_t result = coeff * expo * pow(val, expo - 1);
 
     return coeff * expo * pow(val, expo - 1);
 }
 */ 
 
-long long mtpk::Calculus::derivative_term(std::string p_term, long long val) {
+int64_t mtpk::Calculus::derivative_term(std::string p_term, int64_t val) {
 
     printf("DEBUGGING DERIVTERM\n\n"); 
     // Get coefficient
     std::string coeffStr = "";
-    int i;
+    int64_t i;
     
     for (i = 0; p_term[i] != 'x'; i++) {
         coeffStr.push_back(p_term[i]);
     }
-    long long coeff = atol(coeffStr.c_str());
+    int64_t coeff = atol(coeffStr.c_str());
  
     // Get Power (Skip 2 characters for x and ^)
     std::string powStr = "";
@@ -55,21 +55,21 @@ long long mtpk::Calculus::derivative_term(std::string p_term, long long val) {
         powStr.push_back(p_term[i]);
         printf("powStr = %s \n", powStr.c_str());
     }
-    long long expo = atol(powStr.c_str());
+    int64_t expo = atol(powStr.c_str());
  
     // For ax^n, we return anx^(n-1)
     printf("powStr = %s \n", powStr.c_str());
     printf("p_term = %s \n", p_term.c_str());
-    printf("coeff = %lld \n", coeff);
-    printf("Exponent = %lld \n", expo);
+    printf("coeff = %ld \n", coeff);
+    printf("Exponent = %ld \n", expo);
 
-    long long result = coeff * expo * pow(val, expo - 1);
+    int64_t result = coeff * expo * pow(val, expo - 1);
 
     return coeff * expo * pow(val, expo - 1);
 }
 
-long long mtpk::Calculus::deriv_at(std::string& poly, int val) {
-    long long ans = 0;
+int64_t mtpk::Calculus::deriv_at(std::string& poly, int64_t val) {
+    int64_t ans = 0;
  
     // We use istringstream to get input in tokens
     std::istringstream is(poly);
@@ -86,12 +86,12 @@ long long mtpk::Calculus::deriv_at(std::string& poly, int val) {
             ans = (ans + derivative_term(p_term, val));
     }
     printf("\nDEBUGGING DERIV AT\n\n");
-    printf("ans = %lld\n", ans);
+    printf("ans = %ld\n", ans);
     return ans;
 }
  
 std::string mtpk::Calculus::deriv(std::string& poly) {
-    long long ans = 0;
+    int64_t ans = 0;
  
     // We use istringstream to get input in tokens
     std::istringstream is(poly);
