+++ modules/ml/k-foldCV.cpp
@@ -24,24 +24,24 @@ TEST_CASE("KF_CV-testSmallSample5Fold") {
 }
 
 TEST_CASE("KF_CV-testLargeSample10Fold") {
-    vector<int> large_sample;
-    for (int i = 0; i < 1000; i++){
+    vector<int64_t> large_sample;
+    for (int64_t i = 0; i < 1000; i++){
         large_sample.emplace_back(i);
     }
-    KF_CV<int> kF_CV = KF_CV<int>(large_sample, 10, 1);
-    for (int i = 0; i < 10; i++){
+    KF_CV<int64_t> kF_CV = KF_CV<int64_t>(large_sample, 10, 1);
+    for (int64_t i = 0; i < 10; i++){
         REQUIRE(100 == kF_CV.get_test_fold(i).size());
         REQUIRE(900 == kF_CV.get_train_fold(i).size());
     }
 }
 
 TEST_CASE("KFCV-testLargeSample5Fold") {
-    vector<int> large_sample;
-    for (int i = 0; i < 1000; i++){
+    vector<int64_t> large_sample;
+    for (int64_t i = 0; i < 1000; i++){
         large_sample.emplace_back(i);
     }
-    KF_CV<int> kF_CV = KF_CV<int>(large_sample, 5, 1);
-    for (int i = 0; i < 5; i++){
+    KF_CV<int64_t> kF_CV = KF_CV<int64_t>(large_sample, 5, 1);
+    for (int64_t i = 0; i < 5; i++){
         REQUIRE(200 == kF_CV.get_test_fold(i).size());
         REQUIRE(800 == kF_CV.get_train_fold(i).size());
     }
