+++ modules/number_theory/primes.cpp
@@ -19,12 +19,12 @@ mtpk::Basics ba;
 mtpk::Primality prim;
 
 
-long long int mtpk::Primality::mod_pow(long long int base, 
-        long long int exponent,
-        long long int mod) {
+int64_t mtpk::Primality::mod_pow(int64_t base, 
+        int64_t exponent,
+        int64_t mod) {
 
-    long long x = 1;
-    long long y = base;
+    int64_t x = 1;
+    int64_t y = base;
 
     while (exponent > 0) {
         if (exponent & 1) {
@@ -37,12 +37,12 @@ long long int mtpk::Primality::mod_pow(long long int base,
     return x;
 }
 
-bool mtpk::Primality::is_prime(int n) {
+bool mtpk::Primality::is_prime(int64_t n) {
     if (n <= 1)
         return false;
  
     // Check from 2 to n-1
-    for (int iter = 2; iter < n; iter++)
+    for (int64_t iter = 2; iter < n; iter++)
         if (n % iter == 0)
             return false;
  
@@ -52,12 +52,12 @@ bool mtpk::Primality::is_prime(int n) {
 /*
  * determining if a given number is likely to be prime
  */
-bool mtpk::Primality::compute_miller_rabin(int d, int n) {
+bool mtpk::Primality::compute_miller_rabin(int64_t d, int64_t n) {
     // Pick a random number in [2..n-2] Corner cases make sure that n > 4
-    int a = 2 + rand() % (n - 4);
+    int64_t a = 2 + rand() % (n - 4);
  
     // Compute a^d % n
-    int x = mod_pow(a, d, n);
+    int64_t x = mod_pow(a, d, n);
  
     if (x == 1  || x == n - 1)
        return true;
@@ -81,7 +81,7 @@ bool mtpk::Primality::compute_miller_rabin(int d, int n) {
     return false;
 }
 
-bool mtpk::Primality::miller_rabin_prime(int n, int iters) {
+bool mtpk::Primality::miller_rabin_prime(int64_t n, int64_t iters) {
     /*
      * this method will return true if n is prime, iters will determine
      * accuracy
@@ -93,13 +93,13 @@ bool mtpk::Primality::miller_rabin_prime(int n, int iters) {
         return true;
 
     // odd will represent an odd number
-    int odd = n - 1;
+    int64_t odd = n - 1;
     while (odd % 2 == 0) {
         odd /= 2;
     }
 
     // iterate given the iters paramater
-    for (int i = 0; i < iters; i ++) {
+    for (int64_t i = 0; i < iters; i ++) {
         // compute using the Miller-Rabin method
         if (!compute_miller_rabin(odd, n)) {
             return false;
@@ -108,7 +108,10 @@ bool mtpk::Primality::miller_rabin_prime(int n, int iters) {
     return true;
 }
 
-void mtpk::Primality::miller_rabin(int iters, int min_val, int max_val) {
+void mtpk::Primality::miller_rabin(int64_t iters, 
+                                int64_t min_val, 
+                                int64_t max_val) {
+
     std::cout << "Primes between " << min_val << 
                 " and " << max_val << std::endl;
     //int d = iters - 1;
@@ -128,11 +131,11 @@ void mtpk::Primality::miller_rabin(int iters, int min_val, int max_val) {
 /*
  * another algorithm capable of finding primes
  */
-int mtpk::Primality::jacobian_number(long long a, long long n) {
+int64_t mtpk::Primality::jacobian_number(int64_t a, int64_t n) {
     if (!a)
         return 0;// (0/n) = 0
  
-    int ans = 1;
+    int64_t ans = 1;
 
     if (a < 0) {
         a = -a; // (a/n) = (-a/n)*(-1/n)
@@ -176,18 +179,18 @@ int mtpk::Primality::jacobian_number(long long a, long long n) {
  * 
  * a^(p-1)/2 = (a/p) (mod p)
  */
-bool mtpk::Primality::solovoy_strassen(long long p, int iters) {
+bool mtpk::Primality::solovoy_strassen(int64_t p, int64_t iters) {
     if (p < 2)
         return false;
     
     if (p != 2 && p % 2 == 0)
         return false;
  
-    for (int i = 0; i < iters; i++) {
+    for (int64_t i = 0; i < iters; i++) {
         // Generate a random number a
-        long long a = rand() % (p - 1) + 1;
-        long long jacobian = (p + jacobian_number(a, p)) % p;
-        long long mod = mod_pow(a, (p - 1) / 2, p);
+        int64_t a = rand() % (p - 1) + 1;
+        int64_t jacobian = (p + jacobian_number(a, p)) % p;
+        int64_t mod = mod_pow(a, (p - 1) / 2, p);
  
         if (!jacobian || mod != jacobian)
             return false;
@@ -200,8 +203,8 @@ bool mtpk::Primality::solovoy_strassen(long long p, int iters) {
  * are composite integers satisfying the congruence forumla below
  * b^n - 1 = b (mod n)
  */
-bool mtpk::Primality::carmichael_num(int n) {
-    for (int b = 2; b < n; b++) {
+bool mtpk::Primality::carmichael_num(int64_t n) {
+    for (int64_t b = 2; b < n; b++) {
         // If "b" is relatively prime to n
         if (ba.op_gcd(b, n) == 1)
             // And pow(b, n-1)%n is not 1, return false.
@@ -216,27 +219,27 @@ bool mtpk::Primality::carmichael_num(int n) {
  * logical arithmetic operations, the greek algorithm sieve of Eratosthenes
  * is able to capture all prime numbers to any given limit
  */
-void mtpk::Primality::sieve_of_eratosthenes(int n) {
+void mtpk::Primality::sieve_of_eratosthenes(int64_t n) {
     // Create a boolean array "prime[0..n]" and initialize
     // all entries it as true. A value in prime[i] will
     // finally be false if i is Not a prime, else true.
     bool prime[n + 1];
     memset(prime, true, sizeof(prime));
  
-    for (int p = 2; p * p <= n; p++) {
+    for (int64_t p = 2; p * p <= n; p++) {
         // If prime[p] is not changed, then it is a prime
         if (prime[p] == true) {
             // Update all multiples of p greater than or
             // equal to the square of it numbers which are
             // multiple of p and are less than p^2 are
             // already been marked.
-            for (int i = p * p; i <= n; i += p)
+            for (int64_t i = p * p; i <= n; i += p)
                 prime[i] = false;
         }
     }
  
     // Print all prime numbers
-    for (int p = 2; p <= n; p++)
+    for (int64_t p = 2; p <= n; p++)
         if (prime[p])
             std::cout << p << " " << std::endl;
 }
@@ -256,7 +259,7 @@ void mtpk::Primality::sieve_of_eratosthenes(int n) {
  * x_3 = 677
  * x_4 = 598        gcd(698 - 26, 1111) = 11
  */
-long long int mtpk::Primality::pollard_rho(long long int n) {
+int64_t mtpk::Primality::pollard_rho(int64_t n) {
     /* initialize random seed */
     srand (time(NULL));
  
@@ -269,16 +272,16 @@ long long int mtpk::Primality::pollard_rho(long long int n) {
         return 2;
  
     /* we will pick from the range [2, N) */
-    long long int x = (rand() % (n - 2)) + 2;
-    long long int y = x;
+    int64_t x = (rand() % (n - 2)) + 2;
+    int64_t y = x;
  
     /* the constant in f(x).
      * Algorithm can be re-run with a different c
      * if it throws failure for a composite. */
-    long long int c = (rand() % (n - 1)) + 1;
+    int64_t c = (rand() % (n - 1)) + 1;
 
     /* Initialize candidate divisor (or result) */
-    long long int divisor = 1; 
+    int64_t divisor = 1; 
  
     /* until the prime factor isn't obtained.
        If n is prime, return n */
@@ -317,10 +320,10 @@ long long int mtpk::Primality::pollard_rho(long long int n) {
  *          Φ(4) = 2
  *          Φ(5) = 4
  */
-int mtpk::Primality::ETF(unsigned int n) {
-    unsigned int result = 1;
+int64_t mtpk::Primality::ETF(uint64_t n) {
+    uint64_t result = 1;
     
-    for (int index = 2; unsigned(index) < n; index++) {
+    for (int64_t index = 2; uint64_t(index) < n; index++) {
         if (ba.op_gcd(index, n) == 1) {
             result++;
         }
